# simple-api/openapi.yaml
# application/json only
# 2 paths: /hello, /add
# /hello: GET, query parameter name, pattern: "^[a-zA-Z]*$"
# /add: GET, query parameters num1, num2, integer type, required: true. The example of num1 is 10, num2 generate
# /user: POST, no response. Parameter "name" with example "John Doe" and not longer than 20 characters, "address" with example "123 Main St", age with example [18, 21, 25]

openapi: 3.0.0
info:
  title: Simple API
  version: 1.0.0
paths:
  /hello:
    get:
      parameters:
        - name: name
          in: query
          required: true
          schema:
            type: string
            pattern: "^[a-zA-Z]*$"
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          description: Bad Request
    post:
      responses:
        '405':
          description: Method Not Allowed
  /add:
    get:
      parameters:
        - name: num1
          in: query
          required: true
          schema:
            type: integer
            example: 10
        - name: num2
          in: query
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  result:
                    type: integer
        '400':
          description: Bad Request
    post:
      responses:
        '405':
          description: Method Not Allowed
  /details:
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: John Doe
                  maxLength: 20
                address:
                  type: string
                  example: 123 Main St
                age:
                  type: string
                  example: [18, 21, 25]
      responses:
        '200':
          description: OK
        '400':
          description: Bad Request

  /sum:
    get:
      parameters:
        - name: x
          in: query
          required: true
          schema:
            type: integer
            minimum: 1
            maximum: 10
        - name: y
          in: query
          required: true
          schema:
            type: integer
            minimum: 1000000000
            maximum: 2000000000
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  result:
                    type: integer
        '400':
          description: Bad Request